+++ src/components/cocktails/index.js
@@ -1,4 +1,0 @@
-export { default as CocktailCreate } from './CocktailCreate.vue'
-export { default as CocktailEdit } from './CocktailEdit.vue'
-export { default as CocktailSpecific } from './CocktailSpecific.vue'
-export { default as CocktailHome } from './CocktailHome.vue'