+++ src/components/recipes/index.js
@@ -1,0 +1,4 @@
+export { default as RecipeCreate } from './RecipeCreate.vue'
+export { default as RecipeEdit } from './RecipeEdit.vue'
+export { default as RecipeSpecific } from './RecipeSpecific.vue'
+export { default as RecipeHome } from './RecipeHome.vue'